Move `PATCH /api/v1/project/1/job/1/` to `/admin/job/1/` (close #1051)
