+++ CHANGELOG.md
@@ -9,7 +9,7 @@
         * Add new `GET` endpoints `api/v1/job/` and `api/v1/project/{project_id}/workflow/{workflow_id}/job/` (\#969, \#1003).
         * Add new `GET` endpoints `api/v1/dataset/` and `api/v1/workflow/` (\#988, \#1003).
         * Add new `GET` endpoint `api/v1/project/{project_id}/dataset/` (\#993).
-        * Add superuser-only `PATCH /api/v1/job/{job_id}/` endpoint (\#1030).
+        * Add `PATCH /admin/job/{job_id}/` endpoint (\#1030, \#1053).
         * Move `GET /auth/whoami/` to `GET /auth/current-user/` (\#1013).
         * Move `PATCH /auth/users/me/` to `PATCH /auth/current-user/` (\#1013, \#1035).
         * Remove `DELETE /auth/users/{id}/` endpoint (\#994).
