+++ fractal_server/app/routes/admin.py
@@ -6,6 +6,8 @@ from typing import Optional
 
 from fastapi import APIRouter
 from fastapi import Depends
+from fastapi import HTTPException
+from fastapi import status as fastapi_status
 from sqlalchemy import func
 from sqlmodel import select
 
@@ -18,6 +20,7 @@ from ..models import Project
 from ..models import Workflow
 from ..models.security import UserOAuth as User
 from ..schemas import ApplyWorkflowRead
+from ..schemas import ApplyWorkflowUpdate
 from ..schemas import DatasetRead
 from ..schemas import ProjectRead
 from ..schemas import WorkflowRead
@@ -212,3 +215,39 @@ async def view_job(
     await db.close()
 
     return job_list
+
+
+@router_admin.patch(
+    "/job/{job_id}/",
+    response_model=ApplyWorkflowRead,
+)
+async def update_job(
+    job_update: ApplyWorkflowUpdate,
+    job_id: int,
+    user: User = Depends(current_active_superuser),
+    db: AsyncSession = Depends(get_db),
+) -> Optional[ApplyWorkflowRead]:
+    """
+    Change the status of an existing job.
+
+    This endpoint is only open to superusers, and it does not apply
+    project-based access-control to jobs.
+    """
+    job = await db.get(ApplyWorkflow, job_id)
+    if job is None:
+        raise HTTPException(
+            status_code=fastapi_status.HTTP_404_NOT_FOUND,
+            detail=f"Job {job_id} not found",
+        )
+
+    if job_update.status != JobStatusType.FAILED:
+        raise HTTPException(
+            status_code=fastapi_status.HTTP_422_UNPROCESSABLE_ENTITY,
+            detail=f"Cannot set job status to {job_update.status}",
+        )
+
+    setattr(job, "status", job_update.status)
+    await db.commit()
+    await db.refresh(job)
+    await db.close()
+    return job
